+++ couch/tests/conftest.py
@@ -54,9 +54,12 @@ def dd_environment():
     with docker_run(
         compose_file=os.path.join(common.HERE, 'compose', 'compose_v{}.yaml'.format(couch_version)),
         env_vars=env,
-        conditions=[CheckEndpoints([common.URL]), lambda: generate_data(couch_version), lambda: time.sleep(10)],
+        conditions=[CheckEndpoints([common.URL]), lambda: generate_data(couch_version), lambda: time.sleep(20)],
     ):
-        yield common.BASIC_CONFIG
+        if couch_version == '1':
+            yield common.BASIC_CONFIG
+        elif couch_version == '2':
+            yield common.BASIC_CONFIG_V2
 
 
 def generate_data(couch_version):