+++ datadog_checks_dev/datadog_checks/dev/tooling/e2e/docker.py
@@ -19,9 +19,12 @@ from ...utils import path_join
 class DockerInterface(object):
     ENV_TYPE = 'docker'
 
-    def __init__(self, check, env, base_package=None, config=None, metadata=None, agent_build=None, api_key=None):
+    def __init__(
+        self, check, env, base_package=None, config=None, env_vars=None, metadata=None, agent_build=None, api_key=None
+    ):
         self.check = check
         self.env = env
+        self.env_vars = env_vars
         self.base_package = base_package
         self.config = config or {}
         self.metadata = metadata or {}
@@ -116,7 +119,12 @@ class DockerInterface(object):
                 '-v', '{}:{}'.format(self.config_dir, get_agent_conf_dir(self.check, self.agent_version)),
                 # Mount the check directory
                 '-v', '{}:{}'.format(path_join(get_root(), self.check), self.check_mount_dir),
+                # Mount the /proc directory
+                '-v', '/proc:/host/proc',
             ]
+
+            # Any environment variables passed to the start command
+            command.extend('-e {}'.format(var) for var in self.env_vars)
 
             if self.base_package:
                 # Mount the check directory