+++ openldap/tests/conftest.py
@@ -7,7 +7,7 @@ from copy import deepcopy
 import pytest
 
 from datadog_checks.dev import TempDir, docker_run
-from datadog_checks.dev.utils import create_file, file_exists
+from datadog_checks.dev.utils import create_file, path_exists
 from datadog_checks.openldap import OpenLDAP
 from .common import DEFAULT_INSTANCE, HERE, HOST
 
@@ -17,7 +17,7 @@ def dd_environment():
     with TempDir() as d:
         host_socket_path = os.path.join(d, 'ldapi')
 
-        if not file_exists(host_socket_path):
+        if not path_exists(host_socket_path):
             os.chmod(d, 0o770)
             create_file(host_socket_path)
             os.chmod(host_socket_path, 0o640)