+++ postgres/datadog_checks/postgres/postgres.py
@@ -418,6 +418,13 @@ GROUP BY datid, datname
         # Let's use pg8000
         return pg8000.connect, pg8000.InterfaceError, pg8000.ProgrammingError
 
+    def _get_replication_role(self, key, db):
+        cursor = db.cursor()
+        cursor.execute('SELECT pg_is_in_recovery();')
+        role = cursor.fetchone()[0]
+        # value fetched for role is of <type 'bool'>
+        return "standby" if role else "master"
+
     def _get_version(self, key, db):
         if key not in self.versions:
             cursor = db.cursor()
@@ -838,7 +845,7 @@ GROUP BY datid, datname
             "host:%s" % host,
             "port:%s" % port,
         ]
-        service_check_tags.extend(tags)
+        service_check_tags.extend([t for t in tags if not t.startswith('pg_instance')])
         service_check_tags = list(set(service_check_tags))
         return service_check_tags
 
@@ -907,75 +914,77 @@ GROUP BY datid, datname
                 try:
                     self.log.debug("Running query: {}".format(query))
                     cursor.execute(query)
-                    row = cursor.fetchone()
                 except programming_error as e:
                     self.log.error("Error executing query for metric_prefix {}: {}".format(metric_prefix, str(e)))
                     db.rollback()
                     continue
 
-                if not row:
-                    self.log.debug("query result for metric_prefix {}: returned an empty result".format(metric_prefix))
-                    continue
-
-                if len(columns) != len(row):
-                    self.log.error(
-                        "query result for metric_prefix {}: expected {} columns, got {}".format(
-                            metric_prefix, len(columns), len(row)
+                for row in cursor:
+                    if not row:
+                        self.log.debug(
+                            "query result for metric_prefix {}: returned an empty result".format(metric_prefix)
                         )
-                    )
-                    continue
-
-                metric_info = []
-                query_tags = custom_query.get('tags', [])
-                query_tags.extend(tags)
-
-                for column, value in zip(columns, row):
-                    # Columns can be ignored via configuration.
-                    if not column:
                         continue
 
-                    name = column.get('name')
-                    if not name:
+                    if len(columns) != len(row):
                         self.log.error(
-                            "column field `name` is required for metric_prefix `{}`".format(metric_prefix)
+                            "query result for metric_prefix {}: expected {} columns, got {}".format(
+                                metric_prefix, len(columns), len(row)
+                            )
                         )
-                        break
+                        continue
 
-                    column_type = column.get('type')
-                    if not column_type:
-                        self.log.error(
-                            "column field `type` is required for column `{}` "
-                            "of metric_prefix `{}`".format(name, metric_prefix)
-                        )
-                        break
+                    metric_info = []
+                    query_tags = custom_query.get('tags', [])
+                    query_tags.extend(tags)
 
-                    if column_type == 'tag':
-                        query_tags.append('{}:{}'.format(name, value))
+                    for column, value in zip(columns, row):
+                        # Columns can be ignored via configuration.
+                        if not column:
+                            continue
+
+                        name = column.get('name')
+                        if not name:
+                            self.log.error(
+                                "column field `name` is required for metric_prefix `{}`".format(metric_prefix)
+                            )
+                            break
+
+                        column_type = column.get('type')
+                        if not column_type:
+                            self.log.error(
+                                "column field `type` is required for column `{}` "
+                                "of metric_prefix `{}`".format(name, metric_prefix)
+                            )
+                            break
+
+                        if column_type == 'tag':
+                            query_tags.append('{}:{}'.format(name, value))
+                        else:
+                            if not hasattr(self, column_type):
+                                self.log.error(
+                                    "invalid submission method `{}` for column `{}` of "
+                                    "metric_prefix `{}`".format(column_type, name, metric_prefix)
+                                )
+                                break
+                            try:
+                                metric_info.append((
+                                    '{}.{}'.format(metric_prefix, name),
+                                    float(value),
+                                    column_type
+                                ))
+                            except (ValueError, TypeError):
+                                self.log.error(
+                                    "non-numeric value `{}` for metric column `{}` of "
+                                    "metric_prefix `{}`".format(value, name, metric_prefix)
+                                )
+                                break
+
+                    # Only submit metrics if there were absolutely no errors - all or nothing.
                     else:
-                        if not hasattr(self, column_type):
-                            self.log.error(
-                                "invalid submission method `{}` for column `{}` of "
-                                "metric_prefix `{}`".format(column_type, name, metric_prefix)
-                            )
-                            break
-                        try:
-                            metric_info.append((
-                                '{}.{}'.format(metric_prefix, name),
-                                float(value),
-                                column_type
-                            ))
-                        except (ValueError, TypeError):
-                            self.log.error(
-                                "non-numeric value `{}` for metric column `{}` of "
-                                "metric_prefix `{}`".format(value, name, metric_prefix)
-                            )
-                            break
-
-                # Only submit metrics if there were absolutely no errors - all or nothing.
-                else:
-                    for info in metric_info:
-                        metric, value, method = info
-                        getattr(self, method)(metric, value, tags=query_tags)
+                        for info in metric_info:
+                            metric, value, method = info
+                            getattr(self, method)(metric, value, tags=query_tags)
 
     def _get_custom_metrics(self, custom_metrics, key):
         # Pre-processed cached custom_metrics
@@ -1024,6 +1033,7 @@ GROUP BY datid, datname
         collect_activity_metrics = is_affirmative(instance.get('collect_activity_metrics', False))
         collect_database_size_metrics = is_affirmative(instance.get('collect_database_size_metrics', True))
         collect_default_db = is_affirmative(instance.get('collect_default_database', False))
+        tag_replication_role = is_affirmative(instance.get('tag_replication_role', False))
 
         if relations and not dbname:
             self.warning('"dbname" parameter must be set when using the "relations" parameter.')
@@ -1043,6 +1053,9 @@ GROUP BY datid, datname
         else:
             tags = list(set(tags))
 
+        # preset tags to host (if using socket) or host-port
+        tags.extend(["pg_instance:%s" % (host if host.startswith('/') else "{}-{}".format(host, port))])
+
         # preset tags to the database name
         tags.extend(["db:%s" % dbname])
 
@@ -1056,6 +1069,8 @@ GROUP BY datid, datname
             db = self.get_connection(key, host, port, user, password, dbname, ssl, connect_fct, tags)
             version = self._get_version(key, db)
             self.log.debug("Running check against version %s" % version)
+            if tag_replication_role:
+                tags.extend(["replication_role:{}".format(self._get_replication_role(key, db))])
             self._collect_stats(key, db, tags, relations, custom_metrics, collect_function_metrics,
                                 collect_count_metrics, collect_activity_metrics, collect_database_size_metrics,
                                 collect_default_db, interface_error, programming_error)