+++ postgres/tests/test_custom_metrics.py
@@ -23,7 +23,11 @@ def test_custom_metrics(aggregator, pg_instance):
     postgres_check = PostgreSql('postgres', {}, {})
     postgres_check.check(pg_instance)
 
-    aggregator.assert_metric('custom.num', value=21, tags=['customdb:a'] + pg_instance['tags'])
+    tags = ['customdb:a']
+    tags.append('pg_instance:{}-{}'.format(pg_instance['host'], pg_instance['port']))
+    tags.extend(pg_instance['tags'])
+
+    aggregator.assert_metric('custom.num', value=21, tags=tags)
 
 
 @pytest.mark.integration
@@ -33,7 +37,7 @@ def test_custom_queries(aggregator, pg_instance):
         'custom_queries': [
             {
                 'metric_prefix': 'custom',
-                'query': "SELECT letter, num FROM (VALUES (21, 'a'), (22, 'b'), (23, 'c')) AS t (num,letter) LIMIT 1",
+                'query': "SELECT letter, num FROM (VALUES (97, 'a'), (98, 'b'), (99, 'c')) AS t (num,letter)",
                 'columns': [
                     {
                         'name': 'customtag',
@@ -47,7 +51,7 @@ def test_custom_queries(aggregator, pg_instance):
             },
             {
                 'metric_prefix': 'another_custom_one',
-                'query': "SELECT letter, num FROM (VALUES (21, 'a'), (22, 'b'), (23, 'c')) AS t (num,letter) LIMIT 1",
+                'query': "SELECT letter, num FROM (VALUES (97, 'a'), (98, 'b'), (99, 'c')) AS t (num,letter)",
                 'columns': [
                     {
                         'name': 'customtag',
@@ -63,5 +67,14 @@ def test_custom_queries(aggregator, pg_instance):
     })
     postgres_check = PostgreSql('postgres', {}, {})
     postgres_check.check(pg_instance)
-    tags = ['customtag:a', 'db:{}'.format(pg_instance['dbname'])] + pg_instance['tags']
-    aggregator.assert_metric('custom.num', value=21, tags=tags)
+    tags = ['db:{}'.format(pg_instance['dbname'])]
+    tags.append('pg_instance:{}-{}'.format(pg_instance['host'], pg_instance['port']))
+    tags.extend(pg_instance['tags'])
+
+    for tag in ('a', 'b', 'c'):
+        value = ord(tag)
+        custom_tags = ['customtag:{}'.format(tag)]
+        custom_tags.extend(tags)
+
+        aggregator.assert_metric('custom.num', value=value, tags=custom_tags)
+        aggregator.assert_metric('another_custom_one.num', value=value, tags=custom_tags)