+++ snmp/datadog_checks/snmp/snmp.py
@@ -72,6 +72,8 @@ class SnmpCheck(NetworkCheck):
         tags = instance.get("tags", [])
         ip_address = instance["ip_address"]
         metrics = instance.get('metrics', [])
+        if _is_affirmative(instance.get('use_global_metrics', True)):
+            metrics.extend(self.init_config.get('global_metrics', []))
         timeout = int(instance.get('timeout', self.DEFAULT_TIMEOUT))
         retries = int(instance.get('retries', self.DEFAULT_RETRIES))
         enforce_constraints = _is_affirmative(instance.get('enforce_mib_constraints', True))